Force remove secret.properties from tracking and update .gitignore
