+++ Java2Project/src/main/java/Java2Project/dto/arriveBus/ArriveBusItems.java
@@ -1,0 +1,15 @@
+package Java2Project.dto.arriveBus;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+import java.util.List;
+
+@Getter
+@Setter
+@ToString
+public class Items {
+    @JsonFormat(with = JsonFormat.Feature.ACCEPT_SINGLE_VALUE_AS_ARRAY)
+    private List<ItemDto> item;
+}
