+++ Java2Project/src/main/java/Java2Project/dto/busStop/BusStopItemDto.java
@@ -1,0 +1,18 @@
+package Java2Project.dto.busStop;
+
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class ItemDto {
+    private int citycode;
+    private double gpslati;
+    private double gpslong;
+    private String nodeid;
+    private String nodenm;
+    private int nodeno;
+}
