+++ Java2Project/src/main/java/Java2Project/dto/busStop/BusStopItems.java
@@ -1,4 +1,6 @@
-package Java2Project.dto.arriveBus;
+package Java2Project.dto.busStop;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -7,7 +9,7 @@ import java.util.List;
 
 @Getter
 @Setter
-@ToString
 public class Items {
+    @JsonFormat(with = JsonFormat.Feature.ACCEPT_SINGLE_VALUE_AS_ARRAY)
     private List<ItemDto> item;
 }
